- Hace un main chico - Delega resposabilidades de arbol a una clase BtreeP
